[www-mos] kis takarítás illetve tester.html-ben custom üzenetküldés opcio EsbSocket-ben bug, ha újrakapcsolódás közben hívjuk az end() fgv-t akkor connection = false
git-svn-id: svn+ssh://vertigoweb.hu/repo/trunk/www/mesh-owners-site@1198 c8a1e018-962f-4fc7-b34a-25f67a7ad721

+++ start.sh
@@ -8,13 +8,23 @@ node=$(node --version)
 npm=$(npm --version)
 
 echo ""
-echo "********************************************************"
-echo "*        Mesh Data Systems Kft.  -  Owners Site        *"
-echo "*             Using Node version:  $node             *"
-echo "*      Using NPM version: $npm      *"
-echo "********************************************************"
+echo -en '\E[0;32m'"\033[1mMesh Data Systems Kft. :\033[0m" 
+tput sgr0
+echo -e "\033[1m: Owners Site\033[0m"
 echo ""
-echo "Server started..."
+echo -en "\033[1mUsing Node version:\033[0m"
+tput sgr0 
+echo " $node"
+echo -en "\033[1mUsing NPM  version:\033[0m"
+tput sgr0
+echo " v$npm"
 
-killall node
+if [ "$(pidof node)" ] 
+then
+  killall node
+  echo ""
+else
+  echo ""
+fi
+
 node core.js
